feat: sized numeric types and flip the default to 32bit

+++ packages/compiler/runtime/maya_rt.h
@@ -33,28 +33,28 @@ void yuga_maya_plat_window(yuga_str title, int64_t w, int64_t h);
 int64_t yuga_maya_plat_now_ms(void);
 
 /* Generated Yuga (`yuga_maya_engine_*`). */
-int64_t yuga_maya_engine_is_map(void);
-void yuga_maya_engine_set_viewport(int64_t w, int64_t h);
-void yuga_maya_engine_fixed_update(int64_t dt_ms);
+int32_t yuga_maya_engine_is_map(void);
+void yuga_maya_engine_set_viewport(int32_t w, int32_t h);
+void yuga_maya_engine_fixed_update(int32_t dt_ms);
 void yuga_maya_engine_frame(void);
-void yuga_maya_engine_input_key(int64_t key);
-void yuga_maya_engine_cam_orbit(int64_t dx, int64_t dy);
-void yuga_maya_engine_cam_zoom(int64_t dy100);
-void yuga_maya_engine_cam_pan(int64_t dx, int64_t dy);
+void yuga_maya_engine_input_key(int32_t key);
+void yuga_maya_engine_cam_orbit(int32_t dx, int32_t dy);
+void yuga_maya_engine_cam_zoom(int32_t dy100);
+void yuga_maya_engine_cam_pan(int32_t dx, int32_t dy);
 void yuga_maya_engine_cam_reset(void);
-void yuga_maya_engine_note_present(int64_t dt_ms);
-int64_t yuga_maya_engine_sprite_count(void);
-int64_t yuga_maya_engine_sprite_x(int64_t i);
-int64_t yuga_maya_engine_sprite_y(int64_t i);
-int64_t yuga_maya_engine_sprite_r(int64_t i);
-int64_t yuga_maya_engine_sprite_orbit_r(int64_t i);
-int64_t yuga_maya_engine_sprite_rgb(int64_t i);
-int64_t yuga_maya_engine_sprite_is_sun(int64_t i);
-yuga_str yuga_maya_engine_sprite_name(int64_t i);
+void yuga_maya_engine_note_present(int32_t dt_ms);
+int32_t yuga_maya_engine_sprite_count(void);
+int32_t yuga_maya_engine_sprite_x(int32_t i);
+int32_t yuga_maya_engine_sprite_y(int32_t i);
+int32_t yuga_maya_engine_sprite_r(int32_t i);
+int32_t yuga_maya_engine_sprite_orbit_r(int32_t i);
+int32_t yuga_maya_engine_sprite_rgb(int32_t i);
+int32_t yuga_maya_engine_sprite_is_sun(int32_t i);
+yuga_str yuga_maya_engine_sprite_name(int32_t i);
 yuga_str yuga_maya_engine_hud(void);
-int64_t yuga_maya_engine_fb_w(void);
-int64_t yuga_maya_engine_fb_h(void);
-int64_t yuga_maya_engine_fb_at(int64_t i);
+int32_t yuga_maya_engine_fb_w(void);
+int32_t yuga_maya_engine_fb_h(void);
+int32_t yuga_maya_engine_fb_at(int32_t i);
 
 const uint8_t *maya_cpu_fb(int *w, int *h);
