+++ packages/compiler/runtime/zeus_rt.h
@@ -27,7 +27,7 @@ typedef struct {
 /* --- Yuga API: import "std:zeus" → yuga_zeus_* --- */
 
 void yuga_zeus_raw_init(void);
-void yuga_zeus_open_window(yuga_str title, int64_t width, int64_t height);
+void yuga_zeus_open_window(yuga_str title, int32_t width, int32_t height);
 void yuga_zeus_raw_run(void);
 
 Signal yuga_zeus_signal(int64_t value);
@@ -38,24 +38,24 @@ int64_t yuga_zeus_sig_alloc_int(int64_t value);
 int64_t yuga_zeus_sig_alloc_zero(void);
 void yuga_zeus_sig_free(int64_t id);
 void yuga_zeus_hook_begin(void);
-Signal yuga_zeus_hook_signal(int64_t value);
+Signal yuga_zeus_hook_signal(int32_t value);
 int64_t yuga_zeus_get(Signal sig);
 void yuga_zeus_set(Signal sig, int64_t value);
-void yuga_zeus_inc(Signal sig, int64_t delta);
+void yuga_zeus_inc(Signal sig, int32_t delta);
 
 void yuga_zeus_sig_bind(int64_t id, const void *src, int64_t n);
 void yuga_zeus_sig_load(int64_t id, void *dst, int64_t n);
 int64_t yuga_zeus_sig_changed(int64_t id, const void *src, int64_t n);
 int64_t yuga_zeus_sig_gen(int64_t id);
 Signal yuga_zeus_appearance(void);
-void yuga_zeus_theme(int64_t slot, int64_t light, int64_t dark);
-int64_t yuga_zeus_role(int64_t slot);
+void yuga_zeus_theme(int32_t slot, int32_t light, int32_t dark);
+int32_t yuga_zeus_role(int32_t slot);
 
 Node yuga_zeus_raw_row(void);
 Node yuga_zeus_raw_col(void);
 Node yuga_zeus_raw_box(void);
-Node yuga_zeus_grid(int64_t columns, int64_t gap);
-Node yuga_zeus_grid_auto(int64_t min_width, int64_t gap);
+Node yuga_zeus_grid(int32_t columns, int32_t gap);
+Node yuga_zeus_grid_auto(int32_t min_width, int32_t gap);
 Node yuga_zeus_raw_label(yuga_str text);
 Node yuga_zeus_raw_button(yuga_str text);
 Node yuga_zeus_raw_spacer(void);
@@ -71,68 +71,68 @@ Node yuga_zeus_raw_svg(yuga_str markup);
 Node yuga_zeus_child(Node parent, Node node);
 Node yuga_zeus_root(Node node);
 
-Node yuga_zeus_grow(Node node, int64_t weight);
-Node yuga_zeus_shrink(Node node, int64_t weight);
-Node yuga_zeus_pad(Node node, int64_t all);
-Node yuga_zeus_padX(Node node, int64_t px);
-Node yuga_zeus_padY(Node node, int64_t py);
-Node yuga_zeus_padTop(Node node, int64_t px);
-Node yuga_zeus_padRight(Node node, int64_t px);
-Node yuga_zeus_padBottom(Node node, int64_t px);
-Node yuga_zeus_padLeft(Node node, int64_t px);
-Node yuga_zeus_margin(Node node, int64_t all);
-Node yuga_zeus_marginX(Node node, int64_t px);
-Node yuga_zeus_marginY(Node node, int64_t py);
-Node yuga_zeus_marginTop(Node node, int64_t px);
-Node yuga_zeus_marginRight(Node node, int64_t px);
-Node yuga_zeus_marginBottom(Node node, int64_t px);
-Node yuga_zeus_marginLeft(Node node, int64_t px);
-Node yuga_zeus_border(Node node, int64_t width);
-Node yuga_zeus_border_color(Node node, int64_t rgb);
-Node yuga_zeus_gap(Node node, int64_t g);
-Node yuga_zeus_gap_row(Node node, int64_t g);
-Node yuga_zeus_gap_col(Node node, int64_t g);
-Node yuga_zeus_bg(Node node, int64_t rgb);
-Node yuga_zeus_fg(Node node, int64_t rgb);
-Node yuga_zeus_w(Node node, int64_t px);
-Node yuga_zeus_h(Node node, int64_t px);
-Node yuga_zeus_width_pct(Node node, int64_t pct);
-Node yuga_zeus_height_pct(Node node, int64_t pct);
-Node yuga_zeus_min_w(Node node, int64_t px);
-Node yuga_zeus_max_w(Node node, int64_t px);
-Node yuga_zeus_min_h(Node node, int64_t px);
-Node yuga_zeus_max_h(Node node, int64_t px);
-Node yuga_zeus_aspect(Node node, int64_t aw, int64_t ah);
+Node yuga_zeus_grow(Node node, int32_t weight);
+Node yuga_zeus_shrink(Node node, int32_t weight);
+Node yuga_zeus_pad(Node node, int32_t all);
+Node yuga_zeus_padX(Node node, int32_t px);
+Node yuga_zeus_padY(Node node, int32_t py);
+Node yuga_zeus_padTop(Node node, int32_t px);
+Node yuga_zeus_padRight(Node node, int32_t px);
+Node yuga_zeus_padBottom(Node node, int32_t px);
+Node yuga_zeus_padLeft(Node node, int32_t px);
+Node yuga_zeus_margin(Node node, int32_t all);
+Node yuga_zeus_marginX(Node node, int32_t px);
+Node yuga_zeus_marginY(Node node, int32_t py);
+Node yuga_zeus_marginTop(Node node, int32_t px);
+Node yuga_zeus_marginRight(Node node, int32_t px);
+Node yuga_zeus_marginBottom(Node node, int32_t px);
+Node yuga_zeus_marginLeft(Node node, int32_t px);
+Node yuga_zeus_border(Node node, int32_t width);
+Node yuga_zeus_border_color(Node node, int32_t rgb);
+Node yuga_zeus_gap(Node node, int32_t g);
+Node yuga_zeus_gap_row(Node node, int32_t g);
+Node yuga_zeus_gap_col(Node node, int32_t g);
+Node yuga_zeus_bg(Node node, int32_t rgb);
+Node yuga_zeus_fg(Node node, int32_t rgb);
+Node yuga_zeus_w(Node node, int32_t px);
+Node yuga_zeus_h(Node node, int32_t px);
+Node yuga_zeus_width_pct(Node node, int32_t pct);
+Node yuga_zeus_height_pct(Node node, int32_t pct);
+Node yuga_zeus_min_w(Node node, int32_t px);
+Node yuga_zeus_max_w(Node node, int32_t px);
+Node yuga_zeus_min_h(Node node, int32_t px);
+Node yuga_zeus_max_h(Node node, int32_t px);
+Node yuga_zeus_aspect(Node node, int32_t aw, int32_t ah);
 Node yuga_zeus_flex_row(Node node);
 Node yuga_zeus_flex_col(Node node);
 Node yuga_zeus_flex_row_reverse(Node node);
 Node yuga_zeus_flex_col_reverse(Node node);
 Node yuga_zeus_flex_wrap(Node node);
-Node yuga_zeus_align_self(Node node, int64_t mode);
-Node yuga_zeus_span(Node node, int64_t columns);
-Node yuga_zeus_radius(Node node, int64_t px);
-Node yuga_zeus_justify(Node node, int64_t mode);
-Node yuga_zeus_align(Node node, int64_t mode);
-Node yuga_zeus_position(Node node, int64_t mode);
-Node yuga_zeus_top(Node node, int64_t px);
-Node yuga_zeus_right(Node node, int64_t px);
-Node yuga_zeus_bottom(Node node, int64_t px);
-Node yuga_zeus_left(Node node, int64_t px);
-Node yuga_zeus_z_index(Node node, int64_t z);
-Node yuga_zeus_opacity(Node node, int64_t pct);
-Node yuga_zeus_overflow(Node node, int64_t mode);
-Node yuga_zeus_font(Node node, int64_t px);
+Node yuga_zeus_align_self(Node node, int32_t mode);
+Node yuga_zeus_span(Node node, int32_t columns);
+Node yuga_zeus_radius(Node node, int32_t px);
+Node yuga_zeus_justify(Node node, int32_t mode);
+Node yuga_zeus_align(Node node, int32_t mode);
+Node yuga_zeus_position(Node node, int32_t mode);
+Node yuga_zeus_top(Node node, int32_t px);
+Node yuga_zeus_right(Node node, int32_t px);
+Node yuga_zeus_bottom(Node node, int32_t px);
+Node yuga_zeus_left(Node node, int32_t px);
+Node yuga_zeus_z_index(Node node, int32_t z);
+Node yuga_zeus_opacity(Node node, int32_t pct);
+Node yuga_zeus_overflow(Node node, int32_t mode);
+Node yuga_zeus_font(Node node, int32_t px);
 
 Node yuga_zeus_bind(Node label, Signal sig);
-Node yuga_zeus_bind_n(Node label, int64_t n);
-Node yuga_zeus_digits(int64_t n);
+Node yuga_zeus_bind_n(Node label, int32_t n);
+Node yuga_zeus_digits(int32_t n);
 Node yuga_zeus_on_click_inc(Node button, Signal sig);
 Node yuga_zeus_on_click_toggle(Node node, Signal sig);
-Node yuga_zeus_on_click_set(Node node, Signal sig, int64_t value);
-Node yuga_zeus_on_click_add(Node node, Signal sig, int64_t delta);
+Node yuga_zeus_on_click_set(Node node, Signal sig, int32_t value);
+Node yuga_zeus_on_click_add(Node node, Signal sig, int32_t delta);
 Node yuga_zeus_show(Node node, Signal sig);
-Node yuga_zeus_show_eq(Node node, Signal sig, int64_t value);
-Node yuga_zeus_show_ne(Node node, Signal sig, int64_t value);
+Node yuga_zeus_show_eq(Node node, Signal sig, int32_t value);
+Node yuga_zeus_show_ne(Node node, Signal sig, int32_t value);
 Node yuga_zeus_key_context(Node node, yuga_str name);
 Node yuga_zeus_focusable(Node node);
 Node yuga_zeus_capture_text(Node node);
@@ -145,14 +145,14 @@ void yuga_zeus_map_key(yuga_str spec, yuga_str action, yuga_str ctx);
 void yuga_zeus_remap_key(yuga_str spec, yuga_str action, yuga_str ctx);
 int zeus_handle_key_ev(int key, int mods);
 
-Node yuga_zeus_show_ge(Node node, Signal sig, int64_t value);
-Node yuga_zeus_show_le(Node node, Signal sig, int64_t value);
+Node yuga_zeus_show_ge(Node node, Signal sig, int32_t value);
+Node yuga_zeus_show_le(Node node, Signal sig, int32_t value);
 Node yuga_zeus_raw_hover(Node node, Signal sig);
-Node yuga_zeus_hover_delay(Node node, int64_t ms);
-Node yuga_zeus_hover_leave(Node node, int64_t ms);
+Node yuga_zeus_hover_delay(Node node, int32_t ms);
+Node yuga_zeus_hover_leave(Node node, int32_t ms);
 Node yuga_zeus_pulse(Node node);
 Node yuga_zeus_dismiss(Node node);
-Node yuga_zeus_keys(Node node, Signal sig, int64_t lo, int64_t hi);
+Node yuga_zeus_keys(Node node, Signal sig, int32_t lo, int32_t hi);
 Node yuga_zeus_keys_page(Node node, Signal sig);
 
 /* --- Platform hosts: zeus/desktop/mac.m, zeus/ios/ios.m, zeus/web/wasm.c --- */
@@ -213,9 +213,9 @@ int64_t yuga_platform_plat_load_font(yuga_str family, yuga_str src);
 void yuga_zeus_plat_text_rot(int64_t x, int64_t y, yuga_str s, int64_t rgb, int64_t font,
                              int64_t deg);
 void yuga_zeus_plat_text_int(int64_t x, int64_t y, int64_t v, int64_t rgb, int64_t font);
-void yuga_zeus_plat_measure(yuga_str s, int64_t px, int64_t *w, int64_t *h);
-void yuga_zeus_plat_measure_int(int64_t v, int64_t px, int64_t *w, int64_t *h);
-void yuga_zeus_plat_measure_wrap(yuga_str s, int64_t px, int64_t max_w, int64_t *w, int64_t *h);
+void yuga_zeus_plat_measure(yuga_str s, int32_t px, int32_t *w, int32_t *h);
+void yuga_zeus_plat_measure_int(int32_t v, int32_t px, int32_t *w, int32_t *h);
+void yuga_zeus_plat_measure_wrap(yuga_str s, int32_t px, int32_t max_w, int32_t *w, int32_t *h);
 void yuga_zeus_plat_text_wrap(int64_t x, int64_t y, yuga_str s, int64_t rgb, int64_t font,
                              int64_t max_w);
 void yuga_zeus_plat_set_window(yuga_str title, int64_t width, int64_t height);
@@ -225,7 +225,7 @@ void yuga_zeus_plat_svg(int64_t x, int64_t y, int64_t w, int64_t h, yuga_str mar
                        int64_t rgb, int64_t alpha);
 void yuga_zeus_plat_image(int64_t x, int64_t y, int64_t w, int64_t h, yuga_str src,
                          int64_t radius, int64_t alpha, int64_t fit);
-yuga_str yuga_zeus_plat_pick_image(int64_t *w, int64_t *h);
+yuga_str yuga_zeus_plat_pick_image(int32_t *w, int32_t *h);
 void yuga_zeus_plat_save(void);
 void yuga_zeus_plat_clip(int64_t x, int64_t y, int64_t w, int64_t h);
 void yuga_zeus_plat_restore(void);
@@ -250,10 +250,10 @@ void yuga_platform_plat_text(int64_t x, int64_t y, yuga_str s, int64_t rgb, int6
 void yuga_platform_plat_text_rot(int64_t x, int64_t y, yuga_str s, int64_t rgb, int64_t font,
                                  int64_t deg);
 void yuga_platform_plat_text_int(int64_t x, int64_t y, int64_t v, int64_t rgb, int64_t font);
-void yuga_platform_plat_measure(yuga_str s, int64_t px, int64_t *w, int64_t *h);
-void yuga_platform_plat_measure_int(int64_t v, int64_t px, int64_t *w, int64_t *h);
-void yuga_platform_plat_measure_wrap(yuga_str s, int64_t px, int64_t max_w, int64_t *w,
-                                     int64_t *h);
+void yuga_platform_plat_measure(yuga_str s, int32_t px, int32_t *w, int32_t *h);
+void yuga_platform_plat_measure_int(int32_t v, int32_t px, int32_t *w, int32_t *h);
+void yuga_platform_plat_measure_wrap(yuga_str s, int32_t px, int32_t max_w, int32_t *w,
+                                     int32_t *h);
 void yuga_platform_plat_text_wrap(int64_t x, int64_t y, yuga_str s, int64_t rgb, int64_t font,
                                   int64_t max_w);
 void yuga_platform_plat_set_window(yuga_str title, int64_t width, int64_t height);
@@ -263,7 +263,7 @@ void yuga_platform_plat_svg(int64_t x, int64_t y, int64_t w, int64_t h, yuga_str
                             int64_t rgb, int64_t alpha);
 void yuga_platform_plat_image(int64_t x, int64_t y, int64_t w, int64_t h, yuga_str src,
                               int64_t radius, int64_t alpha, int64_t fit);
-yuga_str yuga_platform_plat_pick_image(int64_t *w, int64_t *h);
+yuga_str yuga_platform_plat_pick_image(int32_t *w, int32_t *h);
 void yuga_platform_plat_save(void);
 int64_t yuga_platform_plat_mem_kb(void);
 void yuga_platform_plat_clip(int64_t x, int64_t y, int64_t w, int64_t h);
@@ -312,33 +312,33 @@ int64_t yuga_platform_plat_inset_bottom(void);
 int64_t yuga_platform_plat_inset_left(void);
 
 /* Yuga engine entry points (std/zeus.yuga). Cocoa trampolines through these. */
-void yuga_zeus_engine_layout(int64_t width, int64_t height);
+void yuga_zeus_engine_layout(int32_t width, int32_t height);
 void yuga_zeus_engine_paint(void);
-int64_t yuga_zeus_engine_step(void);
+int32_t yuga_zeus_engine_step(void);
 /* ms until the next async timer is due (0 = none; -1 = a spawn waits). */
-int64_t yuga_zeus_engine_next_ms(void);
-int64_t yuga_zeus_engine_click(int64_t x, int64_t y);
-int64_t yuga_zeus_engine_scroll(int64_t x, int64_t y, int64_t dx, int64_t dy);
-int64_t yuga_zeus_engine_scroll_step(int64_t x, int64_t y, int64_t dx, int64_t dy);
-int64_t yuga_zeus_engine_drag(int64_t x, int64_t y);
-int64_t yuga_zeus_engine_hover(int64_t x, int64_t y);
+int32_t yuga_zeus_engine_next_ms(void);
+int32_t yuga_zeus_engine_click(int32_t x, int32_t y);
+int32_t yuga_zeus_engine_scroll(int32_t x, int32_t y, int32_t dx, int32_t dy);
+int32_t yuga_zeus_engine_scroll_step(int32_t x, int32_t y, int32_t dx, int32_t dy);
+int32_t yuga_zeus_engine_drag(int32_t x, int32_t y);
+int32_t yuga_zeus_engine_hover(int32_t x, int32_t y);
 void yuga_zeus_engine_mouseup(void);
-int64_t yuga_zeus_engine_over_button(void);
+int32_t yuga_zeus_engine_over_button(void);
 yuga_str yuga_zeus_engine_cursor(void);
-void yuga_zeus_engine_key_apply(int64_t sig, int64_t mode, int64_t value, int64_t lo,
-                               int64_t hi);
+void yuga_zeus_engine_key_apply(int32_t sig, int32_t mode, int32_t value, int32_t lo,
+                               int32_t hi);
 void yuga_zeus_engine_fill_focus(void);
-int64_t yuga_zeus_engine_focus_depth(void);
-int64_t yuga_zeus_engine_focus_node(int64_t i);
-int64_t yuga_zeus_engine_focus_ctx(int64_t i);
-int64_t yuga_zeus_engine_focus_step(int64_t back);
-int64_t yuga_zeus_engine_focus_captures_text(void);
-int64_t yuga_zeus_engine_key(int64_t key);
-int64_t yuga_zeus_engine_key_up(int64_t key, int64_t mods);
-void yuga_zeus_engine_set_mods(int64_t mods);
-int64_t yuga_zeus_engine_insert(yuga_str text);
-int64_t yuga_zeus_engine_marked(yuga_str text);
-void yuga_zeus_engine_picked_image(yuga_str src, int64_t w, int64_t h);
+int32_t yuga_zeus_engine_focus_depth(void);
+int32_t yuga_zeus_engine_focus_node(int32_t i);
+int32_t yuga_zeus_engine_focus_ctx(int32_t i);
+int32_t yuga_zeus_engine_focus_step(int32_t back);
+int32_t yuga_zeus_engine_focus_captures_text(void);
+int32_t yuga_zeus_engine_key(int32_t key);
+int32_t yuga_zeus_engine_key_up(int32_t key, int32_t mods);
+void yuga_zeus_engine_set_mods(int32_t mods);
+int32_t yuga_zeus_engine_insert(yuga_str text);
+int32_t yuga_zeus_engine_marked(yuga_str text);
+void yuga_zeus_engine_picked_image(yuga_str src, int32_t w, int32_t h);
 
 void zeus_layout(int64_t width, int64_t height);
 int zeus_step(float dt);
